+++ Modules/booking/Model/BkCalendarEntry.php
@@ -109,7 +109,7 @@ class BkCalendarEntry extends Model {
         $resources = $this->runRequest($sql1, array($id_space))->fetchAll();
         
         // get all responsibles
-        $sql2 = "SELECT DISTINCT responsible_id FROM bk_calendar_entry WHERE deleted=0 AND space_id=? AND resource_id IN (SELECT id FROM re_info WHERE id_space=? AND deleted=0) AND start_time>=? AND start_time<=?";
+        $sql2 = "SELECT DISTINCT responsible_id FROM bk_calendar_entry WHERE deleted=0 AND id_space=? AND resource_id IN (SELECT id FROM re_info WHERE id_space=? AND deleted=0) AND start_time>=? AND start_time<=?";
         $resps = $this->runRequest($sql2, array($id_space, $id_space, $dateBeginTime, $dateEndTime))->fetchAll();
         
         
@@ -119,7 +119,7 @@ class BkCalendarEntry extends Model {
         
         foreach($resps as $resp){
             
-            $sqlr = "SELECT name from cl_clients WHERE id=? AND deleted=0 AND space_id=?";
+            $sqlr = "SELECT name from cl_clients WHERE id=? AND deleted=0 AND id_space=?";
             //$sqlr = "SELECT name, firstname FROM core_users WHERE id=?";
 
             $respinfo = $this->runRequest($sqlr, array($resp[0], $id_space))->fetch();
@@ -128,7 +128,7 @@ class BkCalendarEntry extends Model {
             }
             $resourceCount = array();
             foreach( $resources as $resource){
-                $sql3 = "SELECT * FROM bk_calendar_entry WHERE deleted=0 AND space_id=? AND responsible_id=? AND resource_id=? AND start_time>=? AND start_time<=?";
+                $sql3 = "SELECT * FROM bk_calendar_entry WHERE deleted=0 AND id_space=? AND responsible_id=? AND resource_id=? AND start_time>=? AND start_time<=?";
                 $res = $this->runRequest($sql3, array($id_space, $resp[0], $resource["id"], $dateBeginTime, $dateEndTime))->fetchAll();
                 if(!$res) { continue; }
 
@@ -147,7 +147,7 @@ class BkCalendarEntry extends Model {
 
     public function updateNullResponsibles($id_space) {
 
-        $sql = "SELECT * FROM bk_calendar_entry WHERE responsible_id<=1 AND deleted=0 AND space_id=?";
+        $sql = "SELECT * FROM bk_calendar_entry WHERE responsible_id<=1 AND deleted=0 AND id_space=?";
         $data = $this->runRequest($sql, array($id_space))->fetchAll();
         $modelUserClient = new ClClientUser();
         
@@ -635,7 +635,7 @@ class BkCalendarEntry extends Model {
      * @return multitype:
      */
     public function selectEntriesByDescription($id_space, $desciption) {
-        $sql = "SELECT * FROM bk_calendar_entry WHERE short_description=? AND space_id=? AND deleted=0 ORDER BY end_time";
+        $sql = "SELECT * FROM bk_calendar_entry WHERE short_description=? AND id_space=? AND deleted=0 ORDER BY end_time";
         $req = $this->runRequest($sql, array($desciption, $id_space));
         return $req->fetchAll();
     }