+++ CHANGELOG.md
@@ -14,6 +14,9 @@
 ## 2.1.9
 
 * [users] fix users_info sql
+* [helpdesk] check if message is an auto-reply and log
+* [booking][invoice] fix sql
+* [booking][colorcode] fix sql
 
 ## 2.1.8