+++ CHANGELOG.md
@@ -8,6 +8,17 @@
 * [stats] count number of tickets per status
 * [stats] add stat calentry_cancel on booking cancel
 * [mail] allow users to unsubscribe to notifications #382
+* [helpdesk] ignore delivery status notifications (do not reply)
+* [helpdesk] let user select multiple tickets to spam them #393
+
+## 2.1.9
+
+* [users] fix users_info sql
+* [helpdesk] check if message is an auto-reply and log
+* [booking][invoice] fix sql
+* [booking][colorcode] fix sql
+* [sql] fix wrong space_id error => id_space (param and sql)
+* [booking][calsup] fix remove unlisted supinfo
 
 ## 2.1.8